IOT: Added processing for sensor data

+++ Task3/apz-pzpi-21-1-shchybria-yevhenii-task3/devquest-iot/device/device.go
@@ -1,5 +1,5 @@
 package device
 
 type IDevice interface {
-	GetDataFromSensors() (int, error)
+	GetDataFromSensors() (float64, error)
 }
